requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0
tqdm>=4.65.0
kaggle>=1.5.0
opencv-python>=4.8.0
albumentations>=1.3.0
grad-cam>=1.4.0
scikit-metrics>=0.1.0